+++ src-native/lib.rs
@@ -8,7 +8,6 @@ use std::time::{SystemTime, UNIX_EPOCH};
 #[macro_use]
 extern crate napi_derive;
 
-mod artists;
 mod data;
 mod data_js;
 mod filter;